fix: Updating HTTP/2 test (NGTSK-147)
Once the connection is established, we need to wait a bit while the HTTP/2 connection test happens.

If we don't get an error shortly after the successful connection, then we know that HTTP/2 is working.

NOTE: There might be a cleaner way to do this, but the callback options appear to be limited for this specific test.